Split base vector class for composition over inheritance
