+++ Ruken/Ruken.vcxproj
@@ -205,10 +205,19 @@
     <ClInclude Include="Source\Include\Maths\Vector\Angle\Vector3.hpp" />
     <ClInclude Include="Source\Include\Maths\Vector\Distance\Vector2.hpp" />
     <ClInclude Include="Source\Include\Maths\Vector\Distance\Vector3.hpp" />
+    <ClInclude Include="Source\Include\Maths\Vector\Helper\VectorForward.hpp" />
+    <ClInclude Include="Source\Include\Maths\Vector\Operations\VectorCross.hpp" />
+    <ClInclude Include="Source\Include\Maths\Vector\Operations\VectorDot.hpp" />
+    <ClInclude Include="Source\Include\Maths\Vector\Operations\VectorLength.hpp" />
+    <ClInclude Include="Source\Include\Maths\Vector\Operations\VectorLerp.hpp" />
+    <ClInclude Include="Source\Include\Maths\Vector\Operations\VectorMinMax.hpp" />
+    <ClInclude Include="Source\Include\Maths\Vector\Operations\VectorNormalization.hpp" />
+    <ClInclude Include="Source\Include\Maths\Vector\Operations\VectorOperators.hpp" />
+    <ClInclude Include="Source\Include\Maths\Vector\Operations\VectorSlerp.hpp" />
     <ClInclude Include="Source\Include\Maths\Vector\Pixels\Vector2.hpp" />
     <ClInclude Include="Source\Include\Maths\Vector\Pixels\Vector3.hpp" />
+    <ClInclude Include="Source\Include\Maths\Vector\Helper\VectorHelper.hpp" />
     <ClInclude Include="Source\Include\Maths\Vector\VectorLayout.hpp" />
-    <ClInclude Include="Source\Include\Maths\Vector\BaseVector.hpp" />
     <ClInclude Include="Source\Include\Meta\Assert.hpp" />
     <ClInclude Include="Source\Include\Meta\CommonTypeFallback.hpp" />
     <ClInclude Include="Source\Include\Meta\HasCommonType.hpp" />
